# TaskPro_BackEnd_v0.0 backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "TaskPro Backend API",
    "description": "Professional Task Management REST API with JWT authentication, MongoDB Atlas integration, Cloudinary file storage, and enterprise security features. Containerized with Docker for seamless deployment with real cloud database connectivity.",
    "contact": {
      "name": "Florin Racila",
      "email": "florin.racila@outlook.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Docker containerized deployment (connects to MongoDB Atlas)"
    },
    {
      "url": "https://taskpro-api.herokuapp.com",
      "description": "Production cloud deployment"
    }
  ],
...
  "paths": {
...
    "/api/users/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User registration",
        "description": "Register a new user account with email confirmation. Rate limited to 5 requests per 15 minutes per IP.",
        "parameters": [],
        "requestBody": {
          "description": "User registration data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful, confirmation email sent",
            "headers": {
              "RateLimit-Limit": {
                "description": "Request limit per window",
                "schema": {
                  "type": "integer",
                  "example": 5
                }
              },
              "RateLimit-Remaining": {
                "description": "Remaining requests in current window",
                "schema": {
                  "type": "integer",
                  "example": 4
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          },
          "409": {
            "description": "Email already exists",
            "content": {}
          },
          "429": {
            "description": "Too many registration attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "description": "Authenticate user and receive access/refresh tokens. Rate limited to 5 requests per 15 minutes per IP.",
        "parameters": [],
        "requestBody": {
          "description": "Login credentials",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "headers": {
              "RateLimit-Limit": {
                "description": "Request limit per window",
                "schema": {
                  "type": "integer",
                  "example": 5
                }
              },
              "RateLimit-Remaining": {
                "description": "Remaining requests in current window",
                "schema": {
                  "type": "integer",
                  "example": 4
                }
              },
              "RateLimit-Reset": {
                "description": "Window reset time in seconds",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          },
          "401": {
            "description": "Invalid credentials",
            "content": {}
          },
          "429": {
            "description": "Too many authentication attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboard": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get all dashboards for current user",
        "parameters": [],
        "requestBody": {
          "description": "Getting all dashboards",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllDashboardsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllDashboardsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          },
          "401": {
            "description": "Not authorized",
            "content": {}
          }
        }
      },
      "post": {
        "tags": ["Dashboard"],
        "summary": "adding new dashboard",
        "parameters": [],
        "requestBody": {
          "description": "add new dashboard",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNewDashboardRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNewDashboardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          },
          "401": {
            "description": "Not authorized",
            "content": {}
          },
          "409": {
            "description": "Provided dashboard already exist",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegistrationRequest": {
        "type": "object",
        "required": ["email", "password", "name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "User's name"
          },
          "email": {
            "type": "string",
            "description": "User's email",
            "format": "email@email.com"
          },
          "password": {
            "type": "string",
            "description": "User's password",
            "example": "qwerty123"
          }
        }
      },
      "RegistrationResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User's email",
            "format": "email"
          }
        },
        "example": [{ "email": "1@gmail.com" }]
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "description": "email@email.com" },
          "password": { "type": "string", "description": "secret-password" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "token",
            "example": "qwerty123"
          },
          "user": {
            "type": "object",
            "properties": {
              "_id": { "type": "string", "description": "DB user id" },
              "email": { "type": "string", "description": "email@email.com" },
              "name": {
                "type": "string",
                "description": "name",
                "example": "user-name"
              },
              "avatarURL": {
                "type": "string",
                "description": "avatar address on cloudinary"
              },
              "theme": {
                "type": "string",
                "description": "theme",
                "example": "violet"
              }
            }
          }
        }
      },
...
      "GetAllDashboardsRequest": {},
      "GetAllDashboardsResponse": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "description": "id from DB",
              "example": "viodghfj5435346yuletsd"
            },
            "name": {
              "type": "string",
              "description": "name",
              "example": "User-name"
            },
            "backgroundURL": {
              "type": "string",
              "description": "URL"
            },
            "icon": {
              "type": "string",
              "description": "URL"
            },
            "owner": {
              "type": "string",
              "description": "owner",
              "example": "DB-user-ID"
            }
          }
        }
      },
      "AddNewDashboardResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "id from DB",
            "example": "viodghfj5435346yuletsd"
          },
          "backgroundURL": { "type": "string", "description": "backgroundURL" },
          "owner": { "type": "string", "description": "owner" },
          "name": {
            "type": "string",
            "description": "name",
            "example": "user-name"
          }
        }
      },
...
      "RateLimitError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "🚫 Too many requests from this IP"
          },
          "message": {
            "type": "string",
            "example": "Please try again after 15 minutes"
          },
          "retryAfter": {
            "type": "string",
            "example": "15 minutes"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}